+++ BS440.ini
@@ -5,9 +5,11 @@
 
 
 # Scale properties
+# Model: BS410, BS430, BS440, BS444
 [Scale] 
 ble_address: aa:bb:cc:11:22:33
 device_name: 0202B6332211CCBBAA
+device_model: BS440
 
 # Program behaviour
 # Loglevel: debug | info | critical | error
